Add stop words removal support
Stop words removal support is introduced, as well as a configuration
file (config.json) through which the user can choose if he wants to
remove stop words or not.

To support the additional configuration file, the Config class is
introduced in the config directory. It reads the data from the
config.json file and stores it in a dictionary that is easily accesible
by the program.